You can run tests now (#63)
Added support for running tests.  WDLs and inputs are assembled and submitted to Cromwell to run, and run record is stored in DB.  There is also a new module `status-manager` which runs in a separate thread and queries cromwell for changes in status of running tests.  When statuses change, they are updated in the DB, and if the run finishes, outputs are stored (or referenced, in the case of files) in the DB.