Jupyter notebook reports (#118)
Added support for generating reports from completed test runs.  Reports are defined in the form of Jupyter Notebooks via a new entity called Report.  Once a report is defined, it can be mapped to a template, and then filled reports will be generated automatically for every completed test run based on that template.  They can also be generated manually via the REST API.  Reports are filled by first filling in the data for the run at the top of the notebook and then running the notebook in a WDL.

Co-authored-by: Jonn Smith <jonn-smith@users.noreply.github.com>