Local WDL storage and downloading (#142)
Added a step in template creation and updating to download and store wdls locally. Added REST mappings for retrieving the WDLs for a template. Modified what is returned by calls to Template REST mappings to display the REST mapping for a WDL if it is stored locally, as opposed to the local file path. For WDLs that are currently still referenced in the TEMPLATE table with a gs, http, or https uri, that uri will still be returned
Also, added steps in WDL storage to hash wdl contents and store the hash so we can check against it and avoid redundant files in our WDL storage