Refactored the way config works to no longer be global and to use yaml as the expected format.  (#159)
Refactored the way config works to no longer be global and to use yaml as the expected format.  For this change to work, now a number of the managers, requesters, and other modules previously structured procedurally are now more of an object-oriented structure, with relevant config information and necessary structs being stored as fields within structs with methods for their functionality instead of functions.